Add design proposal for accelerator monitoring.
This was discussed on this doc: https://docs.google.com/document/d/13O4HNrB7QFpKQcLcJm28R-QBH3Xo0VmJ7w_Pkvmsf68/edit
(accessible to anyone who is a member of kubernetes-dev or kubernetes-users Google Groups).

This was presented during the sig-node meeting on 2017-10-10.

Adding this to this repo, so it's more accessible.